chore: refactor dependency track postprocess to prepare for new method
